+++ dropindicator/src/main/java/comulez/github/droplibrary/Touchable.java
@@ -1,6 +1,7 @@
-package comulez.github.dropindicator;
+package comulez.github.droplibrary;
 
-/** Created by eado on 2017/2/23.
+/**
+ * Created by eado on 2017/2/23.
  */
 public interface Touchable {
   void setTouchable(boolean touchable);
