+++ entry/src/main/config.json
@@ -14,8 +14,8 @@
   },
   "deviceConfig": {},
   "module": {
-    "package": "comulez.github.dropindicator",
-    "name": "comulez.github.dropindicatorsample.MyApplication",
+    "package": "comulez.github.droplibrary",
+    "name": "comulez.github.dropindicator.MyApplication",
     "deviceType": [
       "phone"
     ],
@@ -37,7 +37,7 @@
           }
         ],
         "orientation": "unspecified",
-        "name": "comulez.github.dropindicatorsample.MainAbility",
+        "name": "comulez.github.dropindicator.MainAbility",
         "icon": "$media:icon",
         "description": "$string:mainability_description",
         "label": "$string:app_name",
