+++ entry/src/main/java/comulez/github/dropindicator/MyApplication.java
@@ -15,15 +15,11 @@
  */
 package comulez.github.dropindicator;
 
-import ohos.aafwk.ability.delegation.AbilityDelegatorRegistry;
-import org.junit.Test;
+import ohos.aafwk.ability.AbilityPackage;
 
-import static org.junit.Assert.assertEquals;
-
-public class ExampleOhosTest {
-    @Test
-    public void testBundleName() {
-        final String actualBundleName = AbilityDelegatorRegistry.getArguments().getTestBundleName();
-        assertEquals("Validate bundle name","comulez.github.dropindicator", actualBundleName);
-    }
+public class MyApplication extends AbilityPackage {
+  @Override
+  public void onInitialize() {
+    super.onInitialize();
+  }
 }
